refactor(parser): split primary/misc.rs into cohesive submodules (§7-8)
Split the 2216-line parser/primary/misc.rs (over the 500-line limit) into a
facade plus five themed submodules under src/parser/primary/misc/. All but
one are under 500 lines; colonpair.rs (695) holds the single ~530-line
colonpair_expr function, which cannot be split at a function boundary.
misc.rs is free parser functions addressed as misc::<name> (42 call sites);
the facade re-exports each submodule preserving its original visibility, so
no external caller changes. primary/mod.rs is untouched.

- misc.rs (28)        - facade: mod decls, re-exports, ws_inner
- colonpair.rs (695)  - colonpair_expr (single indivisible fn) + sig helpers
- reduction.rs (483)  - reduction_op / reduction_call_style_expr + REDUCTION_OPS
- lambda.rs (481)     - capture_literal / arrow_lambda / block_or_hash / parse_block_body
- hash.rs (392)       - hash-literal body parsing
- anon_decl.rs (189)  - anon class/grammar/role expr + counters

Pure relocation. Moved fns are pub(crate) in submodules and re-exported at
their original visibility; one cross-submodule helper bumped to pub(crate).
Body refs to primary-level siblings rewritten to absolute crate::parser
paths. The fn multiset is byte-identical. No behavior change.
