+++ src/parser/primary/misc.rs
@@ -1,2216 +1,28 @@
-use super::super::parse_result::{PError, PResult, parse_char};
+//! Misc primary parsing: facade re-exporting themed submodules.
+//!
+//! Split from the original monolithic `misc.rs`. Each submodule owns a
+//! cohesive group of free functions; this facade preserves every function's
+//! original visibility through explicit named re-exports so external callers
+//! (`misc::<name>`) keep working unchanged.
 
-use crate::ast::{Expr, Stmt, make_anon_sub};
-use crate::symbol::Symbol;
-use crate::value::Value;
-use crate::value::signature::{make_signature_value, param_defs_to_sig_info};
+mod anon_decl;
+mod colonpair;
+mod hash;
+mod lambda;
+mod reduction;
 
-use super::super::expr::expression;
-use super::super::helpers::{skip_balanced_parens, split_angle_words, ws, ws1};
-use super::super::stmt::keyword;
-use super::parse_call_arg_list;
-use super::string::{double_quoted_string, single_quoted_string};
-use super::var::parse_ident_with_hyphens;
+pub(super) use anon_decl::{anon_class_expr, anon_grammar_expr, anon_role_expr};
+pub(super) use lambda::{arrow_lambda, capture_literal};
+pub(super) use reduction::reduction_op;
 
-use std::collections::HashMap;
-use std::sync::atomic::{AtomicU64, Ordering};
-
-fn skip_pointy_return_type(mut r: &str) -> PResult<'_, Option<String>> {
-    let (r2, _) = ws(r)?;
-    r = r2;
-    if let Some(after_arrow) = r.strip_prefix("-->") {
-        let (after_arrow, type_name) =
-            crate::parser::stmt::parse_return_type_annotation_pub(after_arrow)?;
-        Ok((after_arrow, Some(type_name)))
-    } else {
-        Ok((r, None))
-    }
-}
-
-/// Known reduction operators (must be listed to distinguish from array literals).
-const REDUCTION_OPS: &[&str] = &[
-    "+", "-", "*", "/", "%", "~", "||", "&&", "//", "%%", "**", "^^", "+&", "+|", "+^", "+<", "+>",
-    "~&", "~|", "~^", "~<", "~>", "?&", "?|", "?^", "==", "!=", "<", ">", "<=", ">=", "<=>", "===",
-    "=:=", "!=:=", "=>", "=", "eqv", "eq", "ne", "lt", "gt", "le", "ge", "leg", "cmp", "coll",
-    "unicmp", "~~", "min", "max", "gcd", "lcm", "and", "or", "not", "andthen", "orelse", "xor",
-    ",", "after", "before", "X", "Z", "x", "xx", "&", "|", "^", "o", "∘", "⊍", "div", "mod",
-    "minmax",
-];
-
-/// Find the matching `]` for a `[` at position 0, respecting nesting.
-fn find_matching_bracket(input: &str) -> Option<usize> {
-    let mut depth = 0;
-    for (i, c) in input.char_indices() {
-        match c {
-            '[' => depth += 1,
-            ']' => {
-                depth -= 1;
-                if depth == 0 {
-                    return Some(i);
-                }
-            }
-            _ => {}
-        }
-    }
-    None
-}
-
-/// Flatten nested bracket operator notation into a simple string.
-/// e.g., "+" → "+", "R-" → "R-", "R[+]" → "R+", "[+]" → "+", "R[R[R-]]" → "RRR-"
-fn flatten_bracket_op(s: &str) -> String {
-    if s.is_empty() {
-        return String::new();
-    }
-    let first = s.as_bytes()[0];
-    // Handle plain bracket: [op] → flatten(op)
-    if first == b'['
-        && let Some(end) = find_matching_bracket(s)
-    {
-        let inner = &s[1..end];
-        return flatten_bracket_op(inner);
-    }
-    // Handle meta + bracket: R[op] → R + flatten(op)
-    if (first == b'R' || first == b'Z' || first == b'X')
-        && s.len() > 1
-        && s.as_bytes()[1] == b'['
-        && let Some(end) = find_matching_bracket(&s[1..])
-    {
-        let inner = &s[2..1 + end];
-        let flattened_inner = flatten_bracket_op(inner);
-        let rest = &s[1 + end + 1..];
-        return format!("{}{}{}", first as char, flattened_inner, rest);
-    }
-    s.to_string()
-}
-
-/// Check if the given op (after flattening) is a valid reduction operator.
-/// Handles R/Z/X meta-prefix chains by stripping them to find the base op.
-fn is_valid_reduction_op(op: &str) -> bool {
-    let mut s = op;
-    // Strip meta prefixes while keeping bare operators like `X` intact.
-    while s.len() > 1 {
-        if let Some(rest) = s.strip_prefix('R') {
-            s = rest;
-        } else if let Some(rest) = s.strip_prefix('Z') {
-            s = rest;
-        } else if let Some(rest) = s.strip_prefix('X') {
-            s = rest;
-        } else {
-            break;
-        }
-    }
-    // Also support scan form (\op) and negated operators (!after),
-    // while keeping operators that genuinely start with '!' (e.g. '!=').
-    if let Some(after_scan) = s.strip_prefix('\\') {
-        return is_valid_reduction_op(after_scan);
-    }
-    let s = if s == "∘" { "o" } else { s };
-    if REDUCTION_OPS.contains(&s) {
-        return true;
-    }
-    // Set operators in parenthesized and unicode form
-    if matches!(
-        s,
-        "(-)"
-            | "(|)"
-            | "(&)"
-            | "(.)"
-            | "(^)"
-            | "(elem)"
-            | "(cont)"
-            | "∪"
-            | "∩"
-            | "⊍"
-            | "∖"
-            | "⊖"
-            | "∈"
-            | "∋"
-            | "(<=)"
-            | "(>=)"
-            | "(<)"
-            | "(>)"
-            | "⊆"
-            | "⊇"
-            | "⊂"
-            | "⊃"
-            | "⊄"
-            | "⊅"
-    ) {
-        return true;
-    }
-    if let Some(stripped) = s.strip_prefix('!')
-        && REDUCTION_OPS.contains(&stripped)
-    {
-        return true;
-    }
-    // Hyper operator forms: >>op<<, >>op>>, <<op<<, <<op>>, and Unicode variants
-    if let Some(inner) = strip_hyper_delimiters(s) {
-        return is_valid_reduction_op(inner);
-    }
-    is_custom_reduction_op(s)
-}
-
-/// Strip hyper operator delimiters (>>...<<, >>...>>, <<...<<, <<...>>)
-/// and their Unicode variants, returning the inner operator if found.
-fn strip_hyper_delimiters(s: &str) -> Option<&str> {
-    let after_left = s
-        .strip_prefix(">>")
-        .or_else(|| s.strip_prefix("<<"))
-        .or_else(|| s.strip_prefix('\u{00BB}'))
-        .or_else(|| s.strip_prefix('\u{00AB}'))?;
-    let inner = after_left
-        .strip_suffix(">>")
-        .or_else(|| after_left.strip_suffix("<<"))
-        .or_else(|| after_left.strip_suffix('\u{00BB}'))
-        .or_else(|| after_left.strip_suffix('\u{00AB}'))?;
-    if inner.is_empty() {
-        return None;
-    }
-    Some(inner)
-}
-
-fn is_custom_reduction_op(op: &str) -> bool {
-    if let Some(name) = op.strip_prefix('&') {
-        return is_callable_reduction_name(name);
-    }
-    // Check if it's a user-declared infix symbol operator (e.g. T+, ⋅)
-    if let Some((symbol, len)) =
-        crate::parser::stmt::simple::match_user_declared_infix_symbol_op(op)
-    {
-        if len == op.len() {
-            return true;
-        }
-        // Also accept the symbol as a prefix of op (shouldn't happen for reduction, but be safe)
-        let _ = symbol;
-    }
-    // A bareword that is not a declared infix operator: accept it as a custom
-    // reduction op ONLY if it looks like an operator name (lowercase/`_`-initial).
-    // An uppercase-initial identifier is a type name or class (`[Any]`, `[Int]`,
-    // `[Exception]`), so `[Type]` must parse as an array literal, not a reduction
-    // (`[Any].raku` is `[Any]` array literal's `.raku`, NOT a reduction with op
-    // "Any"). User-declared uppercase infixes are already matched above by
-    // `match_user_declared_infix_symbol_op`.
-    if !op
-        .chars()
-        .next()
-        .is_some_and(|c| c.is_lowercase() || c == '_')
-    {
-        return false;
-    }
-    if let Ok((rest, _)) = parse_ident_with_hyphens(op) {
-        return rest.is_empty();
-    }
-    false
-}
-
-fn is_callable_reduction_name(name: &str) -> bool {
-    if name.is_empty() {
-        return false;
-    }
-    if (name.starts_with("infix:<")
-        || name.starts_with("prefix:<")
-        || name.starts_with("postfix:<")
-        || name.starts_with("term:<"))
-        && name.ends_with('>')
-    {
-        return true;
-    }
-    if let Ok((rest, _)) = parse_ident_with_hyphens(name) {
-        return rest.is_empty();
-    }
-    false
-}
-
-/// Parse a reduction operator: [+], [*], [~], [min], [[+]], [R[+]], etc.
-pub(super) fn reduction_op(input: &str) -> PResult<'_, Expr> {
-    if !input.starts_with('[') {
-        return Err(PError::expected("reduction operator"));
-    }
-    // Use bracket-matching to find the closing ] (supports nesting)
-    let end =
-        find_matching_bracket(input).ok_or_else(|| PError::expected("']' closing reduction"))?;
-    let inner = &input[1..end];
-    if inner.is_empty() {
-        return Err(PError::expected("operator in reduction"));
-    }
-    // Flatten nested bracket notation: [[+]] → "+", [R[+]] → "R+"
-    let op = match flatten_bracket_op(inner).as_str() {
-        "∘" => "o".to_string(),
-        other => other.to_string(),
-    };
-    // A `&callable` reduction operator requires the double-bracket form
-    // `[[&foo]]` (the inner `[&foo]` is itself the bracketed infix). A single
-    // `[&foo]` is an array literal containing the sub, NOT a reduction — in
-    // Raku `[&foo]` alone is a term (array literal), `2 [&foo] 3` is an infix,
-    // and only `[[&foo]] @a` is the reduction. Both forms flatten to op
-    // `&foo`, so distinguish by whether the raw inner was itself bracketed.
-    let meta = |c: char| c == 'R' || c == 'Z' || c == 'X' || c == '\\';
-    let is_callable_amp = |s: &str| {
-        s.strip_prefix('&')
-            .and_then(|rest| rest.chars().next())
-            .is_some_and(|c| c.is_alphabetic() || c == '_')
-    };
-    if is_callable_amp(op.trim_start_matches(meta))
-        && !inner.trim_start().trim_start_matches(meta).starts_with('[')
-    {
-        return Err(PError::expected(
-            "callable reduction requires double-bracket form [[&op]]",
-        ));
-    }
-    // Only accept known operators (after flattening) to avoid confusion with array literals.
-    if !is_valid_reduction_op(&op) {
-        return Err(PError::expected("known reduction operator"));
-    }
-    // Non-associative "structural" infix operators cannot be used in reduction form.
-    // They return Order, not Bool, and are diffy (not chaining).
-    let base_for_check = op.strip_prefix('\\').unwrap_or(op.as_str());
-    let base_for_check = base_for_check
-        .strip_prefix('R')
-        .filter(|s| !s.is_empty())
-        .unwrap_or(base_for_check);
-    if matches!(base_for_check, "leg" | "<=>" | "cmp" | "coll" | "unicmp") {
-        return Err(PError::fatal(format!(
-            "X::Syntax::CannotMeta: Cannot reduce with {base_for_check} because structural infix operators are diffy and not chaining"
-        )));
-    }
-    let r = &input[end + 1..];
-    let call_style_operand = r.starts_with('(');
-    // Zero-argument reduction: [op] with no operands → identity element
-    // Also accept comma-separated context when the operator is clearly a symbolic op
-    // (not an identifier that could be an array element like [Exception]).
-    let is_symbol_op = !op
-        .chars()
-        .next()
-        .is_some_and(|c| c.is_alphabetic() || c == '_');
-    if r.is_empty()
-        || r.starts_with(';')
-        || r.starts_with('}')
-        || r.starts_with(')')
-        || (r.starts_with(',') && is_symbol_op)
-    {
-        return Ok((
-            r,
-            Expr::Reduction {
-                op,
-                expr: Box::new(Expr::ArrayLiteral(vec![])),
-            },
-        ));
-    }
-    // Call-style: [op](args) — use parse_call_arg_list to handle semicolons
-    if call_style_operand {
-        let (r, _) = parse_char(r, '(')?;
-        let (r, _) = ws(r)?;
-        let (r, args) = super::regex::parse_call_arg_list(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = parse_char(r, ')')?;
-        let expr = if args.len() == 1 {
-            args.into_iter().next().unwrap()
-        } else {
-            Expr::ArrayLiteral(args)
-        };
-        return Ok((
-            r,
-            Expr::Reduction {
-                op,
-                expr: Box::new(expr),
-            },
-        ));
-    }
-    // Listop-style reduction (`[+] @a`) requires whitespace between `]`
-    // and the operand for symbolic operators. `[+]@a` would be two terms
-    // in a row. Alphabetic operators like `[min]` are ambiguous with array
-    // literals like `[Exception]` and are not enforced here.
-    if is_symbol_op
-        && !(r.starts_with(' ')
-            || r.starts_with('\t')
-            || r.starts_with('\n')
-            || r.starts_with('\r')
-            || r.starts_with('#'))
-    {
-        return Err(PError::fatal(
-            "X::Syntax::Confused: Two terms in a row".to_string(),
-        ));
-    }
-    let (r, _) = ws(r)?;
-    // Parse comma-separated list as the operand
-    let (r, first) = parse_reduction_operand(r)?;
-    let mut items = vec![first];
-    let mut rest = r;
-    if true {
-        loop {
-            let (r, _) = ws_inner(rest);
-            if !r.starts_with(',') {
-                break;
-            }
-            let r = &r[1..];
-            let (r, _) = ws_inner(r);
-            // Stop at end-of-input, semicolon, closing brackets, or statement modifiers
-            if r.is_empty()
-                || r.starts_with(';')
-                || r.starts_with('}')
-                || r.starts_with(')')
-                || r.starts_with(']')
-            {
-                rest = r;
-                break;
-            }
-            if let Ok((r, next)) = parse_reduction_operand(r) {
-                items.push(next);
-                rest = r;
-            } else {
-                return Err(PError::expected_at(
-                    "expression after ',' in reduction list",
-                    r,
-                ));
-            }
-        }
-    }
-    let mut items = merge_list_infix_seeds(items);
-    items = merge_sequence_seeds(items);
-    let expr = if items.len() == 1 {
-        items.remove(0)
-    } else {
-        Expr::ArrayLiteral(items)
-    };
-    Ok((
-        rest,
-        Expr::Reduction {
-            op,
-            expr: Box::new(expr),
-        },
-    ))
-}
-
-fn parse_reduction_operand(input: &str) -> PResult<'_, Expr> {
-    let (r, _) = ws(input)?;
-    if let Some(r) = r.strip_prefix('|') {
-        let (r, _) = ws(r)?;
-        let (r, expr) = expression(r)?;
-        return Ok((
-            r,
-            Expr::Call {
-                name: Symbol::intern("slip"),
-                args: vec![expr],
-            },
-        ));
-    }
-    expression(r)
-}
-
-fn merge_sequence_seeds(items: Vec<Expr>) -> Vec<Expr> {
-    if items.len() < 2 {
-        return items;
-    }
-    let last = items.last().unwrap();
-    if let Expr::Binary { left, op, right } = last
-        && matches!(
-            op,
-            crate::token_kind::TokenKind::DotDotDot | crate::token_kind::TokenKind::DotDotDotCaret
-        )
-    {
-        let mut seeds: Vec<Expr> = items[..items.len() - 1].to_vec();
-        seeds.push(*left.clone());
-        let merged = Expr::Binary {
-            left: Box::new(Expr::ArrayLiteral(seeds)),
-            op: op.clone(),
-            right: right.clone(),
-        };
-        vec![merged]
-    } else {
-        items
-    }
-}
-
-fn merge_list_infix_seeds(items: Vec<Expr>) -> Vec<Expr> {
-    if items.len() < 2 {
-        return items;
-    }
-    let last = items.last().unwrap();
-    if let Expr::MetaOp {
-        meta,
-        op,
-        left,
-        right,
-    } = last
-        && (meta == "X" || meta == "Z")
-    {
-        let mut seeds: Vec<Expr> = items[..items.len() - 1].to_vec();
-        seeds.push(*left.clone());
-        let merged = Expr::MetaOp {
-            meta: meta.clone(),
-            op: op.clone(),
-            left: Box::new(Expr::ArrayLiteral(seeds)),
-            right: right.clone(),
-        };
-        return vec![merged];
-    }
-    items
-}
-
-pub(in crate::parser) fn reduction_call_style_expr(input: &str) -> PResult<'_, Expr> {
-    if !input.starts_with('[') {
-        return Err(PError::expected("reduction operator"));
-    }
-    let end =
-        find_matching_bracket(input).ok_or_else(|| PError::expected("']' closing reduction"))?;
-    let inner = &input[1..end];
-    if inner.is_empty() {
-        return Err(PError::expected("operator in reduction"));
-    }
-    let op = match flatten_bracket_op(inner).as_str() {
-        "∘" => "o".to_string(),
-        other => other.to_string(),
-    };
-    if !is_valid_reduction_op(&op) {
-        return Err(PError::expected("known reduction operator"));
-    }
-    let r = &input[end + 1..];
-    // Call-style reduction requires `(` immediately after `]` (no whitespace).
-    // `[Z+](args)` is call-style; `[Z+] (a, b), (c, d)` is listop-style
-    // and should be handled by `reduction_op` instead.
-    if !r.starts_with('(') {
-        return Err(PError::expected("call-style reduction operand"));
-    }
-    let rest_after_operand = skip_balanced_parens(r);
-    if rest_after_operand == r {
-        return Err(PError::expected("')' closing reduction argument"));
-    }
-    let close_pos = r.len() - rest_after_operand.len();
-    if close_pos <= 1 {
-        return Err(PError::expected("expression in call-style reduction"));
-    }
-    let inner = &r[1..close_pos - 1];
-    let (inner_rest, args) = parse_call_arg_list(inner)?;
-    let (inner_rest, _) = ws(inner_rest)?;
-    if !inner_rest.is_empty() {
-        return Err(PError::expected("')' in call-style reduction"));
-    }
-    if args.is_empty() {
-        return Err(PError::expected("expression in call-style reduction"));
-    }
-    let inner_expr = if args.len() == 1 {
-        args.into_iter().next().unwrap()
-    } else {
-        Expr::ArrayLiteral(args)
-    };
-    Ok((
-        rest_after_operand,
-        Expr::Reduction {
-            op,
-            expr: Box::new(inner_expr),
-        },
-    ))
-}
-
-/// Parse colonpair expressions: :$var, :@var, :%var, :name(expr), :name, :!name
-pub(in crate::parser) fn colonpair_expr(input: &str) -> PResult<'_, Expr> {
-    let r = input
-        .strip_prefix(':')
-        .filter(|r| !r.starts_with(':'))
-        .ok_or_else(|| PError::expected("colonpair"))?;
-    // :36<...> is a generic radix literal, not a colonpair.
-    let digit_end = r
-        .char_indices()
-        .take_while(|(_, c)| crate::builtins::unicode::unicode_decimal_digit_value(*c).is_some())
-        .last()
-        .map(|(idx, c)| idx + c.len_utf8())
-        .unwrap_or(0);
-    if digit_end > 0 && r[digit_end..].starts_with('<') {
-        return Err(PError::expected("generic radix literal"));
-    }
-    // :2 or :36 alone (digits with no <, (, or [ suffix) is a malformed radix literal.
-    // Only trigger this when the remaining part doesn't start with an identifier char
-    // (to avoid blocking :123name colonpairs).
-    if digit_end > 0 {
-        let after = &r[digit_end..];
-        let next_ch = after.chars().next();
-        let is_ident_follow =
-            next_ch.is_some_and(|c| c.is_alphabetic() || c == '_' || c == '-' || c == '\'');
-        if !is_ident_follow
-            && !after.starts_with('<')
-            && !after.starts_with('(')
-            && !after.starts_with('[')
-        {
-            return Err(PError::expected("generic radix literal"));
-        }
-    }
-    // :{ ... }: typed hash literal (Hash[Mu,Any]).
-    // Unlike bare { ... }, the colon prefix explicitly marks this as a hash,
-    // so we always parse as a hash literal (never as a block).
-    // Object hashes allow non-string keys (e.g., regex keys), so we use
-    // expression-based parsing rather than the simple hash key heuristic.
-    if let Some(after_brace) = r.strip_prefix('{') {
-        let (inner, _) = ws_inner(after_brace);
-        // Empty object hash: :{}
-        if let Some(rest) = inner.strip_prefix('}') {
-            return Ok((rest, Expr::Hash(Vec::new())));
-        }
-        return parse_object_hash_body(inner);
-    }
-    // :(...): signature literal.
-    if let Some(mut r) = r.strip_prefix('(') {
-        let (r2, _) = ws(r)?;
-        // Preferred path: parse using the full parameter list parser.
-        if let Ok((r3, (param_defs, return_type))) =
-            super::super::stmt::parse_param_list_with_return_pub(r2)
-            && let Ok((r3, _)) = ws(r3)
-            && let Ok((r3, _)) = parse_char(r3, ')')
-        {
-            let sig_info = param_defs_to_sig_info(&param_defs, return_type);
-            return Ok((r3, Expr::Literal(make_signature_value(sig_info))));
-        }
-
-        // Fallback path: permissive parsing for legacy forms like :(:$a = True)
-        // used in roast tests around Test::Assuming.
-        r = r2;
-        let mut items = Vec::new();
-        if r.starts_with(')') {
-            let (r, _) = parse_char(r, ')')?;
-            let mut attrs = HashMap::new();
-            attrs.insert("raku".to_string(), Value::str_from(":()"));
-            attrs.insert("perl".to_string(), Value::str_from(":()"));
-            attrs.insert("Str".to_string(), Value::str_from(":()"));
-            attrs.insert("gist".to_string(), Value::str_from(":()"));
-            return Ok((
-                r,
-                Expr::Literal(Value::make_instance(Symbol::intern("Signature"), attrs)),
-            ));
-        }
-        loop {
-            let mut item_input = r;
-            if let Ok((after_type, _type_name)) = parse_ident_with_hyphens(r) {
-                let (after_ws, _) = ws(after_type)?;
-                if after_ws.starts_with(':') {
-                    item_input = after_ws;
-                }
-            }
-            let (r_item, mut item) = match colonpair_expr(item_input) {
-                Ok(parsed) => parsed,
-                Err(_) => match expression(item_input) {
-                    Ok((r_expr, expr_item)) => {
-                        let (r_chk, _) = ws(r_expr)?;
-                        let valid_tail = r_chk.starts_with(',')
-                            || r_chk.starts_with(')')
-                            || keyword("is", r_chk).is_some()
-                            || keyword("where", r_chk).is_some()
-                            || r_chk.starts_with('=');
-                        if valid_tail {
-                            (r_expr, expr_item)
-                        } else {
-                            let (r_frag, frag) = parse_signature_fragment(item_input)?;
-                            (r_frag, Expr::BareWord(frag))
-                        }
-                    }
-                    Err(_) => {
-                        let (r_frag, frag) = parse_signature_fragment(item_input)?;
-                        (r_frag, Expr::BareWord(frag))
-                    }
-                },
-            };
-            let (mut r_next, _) = ws(r_item)?;
-            while let Some(after_is) = keyword("is", r_next) {
-                let (after_is, _) = ws(after_is)?;
-                let (after_is, _trait_name) = parse_ident_with_hyphens(after_is)?;
-                let (after_is, _) = ws(after_is)?;
-                r_next = after_is;
-            }
-            if let Some(after_where) = keyword("where", r_next) {
-                let (after_where, _) = ws(after_where)?;
-                let (after_where, _constraint) = expression(after_where)?;
-                let (after_where, _) = ws(after_where)?;
-                r_next = after_where;
-            }
-            if let Some(after_eq) = r_next.strip_prefix('=')
-                && let Expr::Binary { left, op, .. } = &item
-                && *op == crate::token_kind::TokenKind::FatArrow
-            {
-                let (after_eq, _) = ws(after_eq)?;
-                let (after_eq, value_expr) = expression(after_eq)?;
-                item = Expr::Binary {
-                    left: left.clone(),
-                    op: crate::token_kind::TokenKind::FatArrow,
-                    right: Box::new(value_expr),
-                };
-                let (after_eq, _) = ws(after_eq)?;
-                r_next = after_eq;
-            }
-            items.push(item);
-            if r_next.starts_with(',') {
-                let (r_more, _) = parse_char(r_next, ',')?;
-                let (r_more, _) = ws(r_more)?;
-                r = r_more;
-                continue;
-            }
-            let (r_end, _) = parse_char(r_next, ')')?;
-            let rendered = items
-                .iter()
-                .map(render_signature_item)
-                .collect::<Vec<_>>()
-                .join(", ");
-            let sig = format!(":({})", rendered);
-            let mut attrs = HashMap::new();
-            attrs.insert("raku".to_string(), Value::str(sig.clone()));
-            attrs.insert("perl".to_string(), Value::str(sig.clone()));
-            attrs.insert("Str".to_string(), Value::str(sig.clone()));
-            attrs.insert("gist".to_string(), Value::str(sig));
-            return Ok((
-                r_end,
-                Expr::Literal(Value::make_instance(Symbol::intern("Signature"), attrs)),
-            ));
-        }
-    }
-    // :16(expr) — radix conversion shorthand (equivalent to UNBASE(16, expr))
-    {
-        let mut digit_end = 0;
-        let mut base_clean = String::new();
-        for c in r.chars() {
-            let Some(dv) = crate::builtins::unicode::unicode_decimal_digit_value(c) else {
-                break;
-            };
-            digit_end += c.len_utf8();
-            base_clean.push(char::from_digit(dv, 10).unwrap());
-        }
-        if digit_end > 0 {
-            let after_digits = &r[digit_end..];
-            if let Some(after_paren) = after_digits.strip_prefix('(') {
-                let base: u32 = base_clean.parse().unwrap_or(0);
-                if !(2..=36).contains(&base) {
-                    return Err(PError::expected("generic radix base 2..36"));
-                }
-                let (r_args, _) = ws(after_paren)?;
-                let (r_args, mut args) = parse_call_arg_list(r_args)?;
-                let (r_args, _) = ws(r_args)?;
-                let (r_args, _) = parse_char(r_args, ')')?;
-                let mut call_args = vec![Expr::Literal(Value::Int(base as i64))];
-                call_args.append(&mut args);
-                return Ok((
-                    r_args,
-                    Expr::Call {
-                        name: Symbol::intern("UNBASE"),
-                        args: call_args,
-                    },
-                ));
-            }
-            // :N[list] — radix list notation (equivalent to RADIX_LIST(N, list...))
-            // The list form supports any base >= 2 (no 36 limit since digits are numeric values)
-            if let Some(after_bracket) = after_digits.strip_prefix('[') {
-                let base: u64 = base_clean.parse().unwrap_or(0);
-                if base >= 2 {
-                    let (r_items, _) = ws(after_bracket)?;
-                    let mut items = Vec::new();
-                    let mut r_items = r_items;
-                    while !r_items.starts_with(']') {
-                        let (r2, item) = expression(r_items)?;
-                        items.push(item);
-                        let (r2, _) = ws(r2)?;
-                        if r2.starts_with(',') {
-                            let (r2, _) = parse_char(r2, ',')?;
-                            let (r2, _) = ws(r2)?;
-                            r_items = r2;
-                        } else {
-                            r_items = r2;
-                        }
-                    }
-                    let (r_items, _) = parse_char(r_items, ']')?;
-                    let base_expr = if let Ok(b) = i64::try_from(base) {
-                        Expr::Literal(Value::Int(b))
-                    } else {
-                        Expr::Literal(Value::bigint(num_bigint::BigInt::from(base)))
-                    };
-                    let mut call_args = vec![base_expr];
-                    call_args.extend(items);
-                    return Ok((
-                        r_items,
-                        Expr::Call {
-                            name: Symbol::intern("RADIX_LIST"),
-                            args: call_args,
-                        },
-                    ));
-                }
-            }
-        }
-    }
-    // :123name (numeric leading-value pair) => :name(123)
-    // Handles ASCII digits and Unicode Nd (decimal digit) characters
-    {
-        let mut digit_end = 0;
-        let mut numeric_value: Option<i64> = Some(0);
-        let mut use_bigint = false;
-        for c in r.chars() {
-            let dv = if c.is_ascii_digit() {
-                Some(c as u32 - '0' as u32)
-            } else {
-                crate::builtins::unicode::unicode_decimal_digit_value(c)
-            };
-            if let Some(d) = dv {
-                if let Some(ref mut v) = numeric_value {
-                    match v.checked_mul(10).and_then(|v2| v2.checked_add(d as i64)) {
-                        Some(new_v) => *v = new_v,
-                        None => use_bigint = true,
-                    }
-                }
-                digit_end += c.len_utf8();
-            } else {
-                break;
-            }
-        }
-        if digit_end > 0 && digit_end < r.len() {
-            let digits_str = &r[..digit_end];
-            let after_digits = &r[digit_end..];
-            if let Ok((rest, name)) = parse_ident_with_hyphens(after_digits) {
-                let val = if use_bigint {
-                    // Parse as BigInt for numbers that overflow i64
-                    // Collect decimal digit values and build the number string
-                    let digit_string: String = digits_str
-                        .chars()
-                        .filter_map(|c| {
-                            let dv = if c.is_ascii_digit() {
-                                Some(c as u32 - '0' as u32)
-                            } else {
-                                crate::builtins::unicode::unicode_decimal_digit_value(c)
-                            };
-                            dv.map(|d| char::from_digit(d, 10).unwrap())
-                        })
-                        .collect();
-                    if let Ok(n) = digit_string.parse::<num_bigint::BigInt>() {
-                        Value::bigint(n)
-                    } else {
-                        Value::Int(0)
-                    }
-                } else if numeric_value.is_some() {
-                    Value::Int(numeric_value.unwrap())
-                } else {
-                    Value::Int(0)
-                };
-                // Numeric adverb can't have an extra value: :69th($_) is an error
-                if rest.starts_with('(') {
-                    return Err(PError::expected(
-                        "no extra argument for numeric colonpair (pair already has a numeric value)",
-                    ));
-                }
-                return Ok((
-                    rest,
-                    Expr::Binary {
-                        left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                        op: crate::token_kind::TokenKind::FatArrow,
-                        right: Box::new(Expr::Literal(val)),
-                    },
-                ));
-            }
-        }
-    }
-    // :!name (negated boolean pair)
-    if let Some(after_bang) = r.strip_prefix('!') {
-        let (rest, name) = parse_ident_with_hyphens(after_bang)?;
-        // A negated pair may not carry an argument: `:!foo(3)` is
-        // X::Syntax::NegatedPair ("Argument not allowed on negated pair").
-        if rest.starts_with('(') {
-            let message = format!("Argument not allowed on negated pair with key '{}'", name);
-            let mut attrs = std::collections::HashMap::new();
-            attrs.insert("key".to_string(), Value::str(name.to_string()));
-            attrs.insert("message".to_string(), Value::str(message.clone()));
-            let ex = Value::make_instance(Symbol::intern("X::Syntax::NegatedPair"), attrs);
-            return Err(PError::fatal_with_exception(message, Box::new(ex)));
-        }
-        return Ok((
-            rest,
-            Expr::Binary {
-                left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                op: crate::token_kind::TokenKind::FatArrow,
-                right: Box::new(Expr::Literal(Value::Bool(false))),
-            },
-        ));
-    }
-    // :$var / :@var / :%var / :&var (autopair from variable, with twigil support)
-    if r.starts_with('$') || r.starts_with('@') || r.starts_with('%') || r.starts_with('&') {
-        let sigil = &r[..1];
-        let after_sigil = &r[1..];
-        // Handle $<name> / @<name> / %<name> capture variable twigils
-        if (sigil == "$" || sigil == "@" || sigil == "%")
-            && after_sigil.starts_with('<')
-            && let Some(end) = after_sigil[1..].find('>')
-        {
-            let name = &after_sigil[1..1 + end];
-            let rest = &after_sigil[1 + end + 1..];
-            let var_expr = Expr::CaptureVar(name.to_string());
-            return Ok((
-                rest,
-                Expr::Binary {
-                    left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                    op: crate::token_kind::TokenKind::FatArrow,
-                    right: Box::new(var_expr),
-                },
-            ));
-        }
-        // Check for twigils: *, ?, ^, =, ~, :, !, .
-        let (after_twigil, twigil) = if after_sigil.starts_with('*')
-            || after_sigil.starts_with('?')
-            || after_sigil.starts_with('^')
-            || after_sigil.starts_with('=')
-            || after_sigil.starts_with('~')
-            || after_sigil.starts_with('!')
-            || after_sigil.starts_with('.')
-        {
-            (&after_sigil[1..], &after_sigil[..1])
-        } else if after_sigil.starts_with(':') && !after_sigil.starts_with("::") {
-            (&after_sigil[1..], ":")
-        } else {
-            (after_sigil, "")
-        };
-        let (mut rest, var_name) = parse_ident_with_hyphens(after_twigil)?;
-        // Strip trailing ? or ! for adverb pair forms (:name? :name!)
-        if twigil.is_empty() && (rest.starts_with('?') || rest.starts_with('!')) {
-            rest = &rest[1..];
-        }
-        let key = var_name.to_string();
-        let full_var_name = if twigil.is_empty() {
-            var_name.to_string()
-        } else {
-            format!("{}{}", twigil, var_name)
-        };
-        let var_expr = match sigil {
-            "$" => Expr::Var(full_var_name),
-            "@" => Expr::ArrayVar(full_var_name),
-            "%" => Expr::HashVar(full_var_name),
-            "&" => Expr::CodeVar(full_var_name),
-            _ => unreachable!(),
-        };
-        return Ok((
-            rest,
-            Expr::Binary {
-                left: Box::new(Expr::Literal(Value::str(key))),
-                op: crate::token_kind::TokenKind::FatArrow,
-                right: Box::new(var_expr),
-            },
-        ));
-    }
-    // :name(expr) or :name (boolean true pair)
-    let (mut rest, name) = parse_ident_with_hyphens(r)?;
-    if rest.starts_with('?') || rest.starts_with('!') {
-        rest = &rest[1..];
-    }
-    // Don't parse statement-modifier keywords as bare colonpairs (`:when`),
-    // but allow them when followed by a value (`:when<now>`, `:if(True)`, etc.)
-    if matches!(
-        name,
-        "if" | "unless" | "for" | "while" | "until" | "given" | "when"
-    ) && !rest.starts_with('(')
-        && !rest.starts_with('[')
-        && !rest.starts_with('{')
-        && !rest.starts_with('<')
-        && !rest.starts_with('\u{00AB}')
-    {
-        return Err(PError::expected("colonpair name"));
-    }
-    // Consume unspace between colonpair name and value: :foo\ ("bar")
-    let rest = super::super::helpers::consume_unspace(rest);
-    if let Some(after_paren) = rest.strip_prefix('(') {
-        let (r, _) = ws(after_paren)?;
-        // Handle empty parens: :name() → name => ()
-        if let Some(r) = r.strip_prefix(')') {
-            return Ok((
-                r,
-                Expr::Binary {
-                    left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                    op: crate::token_kind::TokenKind::FatArrow,
-                    right: Box::new(Expr::ArrayLiteral(Vec::new())),
-                },
-            ));
-        }
-        let (r, first) = expression(r)?;
-        let (r, _) = ws(r)?;
-        // Check for separated list: :name(a, b, ...) or :name(a; b; ...) or :name(a,)
-        if r.starts_with(',') || (r.starts_with(';') && !r.starts_with(";;")) {
-            let mut items = vec![first];
-            let mut r = r;
-            while r.starts_with(',') || (r.starts_with(';') && !r.starts_with(";;")) {
-                let sep = if r.starts_with(',') { ',' } else { ';' };
-                let (r2, _) = parse_char(r, sep)?;
-                let (r2, _) = ws(r2)?;
-                // Trailing comma: :name(a,)
-                if r2.starts_with(')') {
-                    r = r2;
-                    break;
-                }
-                let (r2, next) = expression(r2)?;
-                let (r2, _) = ws(r2)?;
-                items.push(next);
-                r = r2;
-            }
-            let (r, _) = parse_char(r, ')')?;
-            return Ok((
-                r,
-                Expr::Binary {
-                    left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                    op: crate::token_kind::TokenKind::FatArrow,
-                    right: Box::new(Expr::ArrayLiteral(items)),
-                },
-            ));
-        }
-        let (r, _) = parse_char(r, ')')?;
-        return Ok((
-            r,
-            Expr::Binary {
-                left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                op: crate::token_kind::TokenKind::FatArrow,
-                right: Box::new(first),
-            },
-        ));
-    }
-    // :name[items] (array-valued colonpair)
-    if rest.starts_with('[') {
-        let (r, _) = parse_char(rest, '[')?;
-        let (r, _) = ws(r)?;
-        let mut items = Vec::new();
-        let mut r = r;
-        while !r.starts_with(']') {
-            let (r2, item) = expression(r)?;
-            items.push(item);
-            let (r2, _) = ws(r2)?;
-            if r2.starts_with(',') {
-                let (r2, _) = parse_char(r2, ',')?;
-                let (r2, _) = ws(r2)?;
-                r = r2;
-            } else {
-                r = r2;
-            }
-        }
-        let (r, _) = parse_char(r, ']')?;
-        return Ok((
-            r,
-            Expr::Binary {
-                left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                op: crate::token_kind::TokenKind::FatArrow,
-                right: Box::new(Expr::BracketArray(items, false)),
-            },
-        ));
-    }
-    // :name{ ... } (block/hash-valued colonpair)
-    if rest.starts_with('{') {
-        let (r, block_or_hash) = block_or_hash_expr(rest)?;
-        return Ok((
-            r,
-            Expr::Binary {
-                left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                op: crate::token_kind::TokenKind::FatArrow,
-                right: Box::new(block_or_hash),
-            },
-        ));
-    }
-    // :name<value> (angle-bracket colonpair, equivalent to :name("value"))
-    if rest.starts_with('<') && !rest.starts_with("<<") && !rest.starts_with("<=") {
-        let inner = &rest[1..];
-        if let Some(close) = super::container::find_nested_angle_close_pub(inner) {
-            let content = &inner[..close];
-            let r = &inner[close + 1..];
-            let words = split_angle_words(content);
-            if !words.is_empty() {
-                let val_expr = if words.len() == 1 {
-                    Expr::Literal(super::container::angle_word_value(words[0]))
-                } else {
-                    Expr::ArrayLiteral(
-                        words
-                            .into_iter()
-                            .map(|w| Expr::Literal(super::container::angle_word_value(w)))
-                            .collect(),
-                    )
-                };
-                return Ok((
-                    r,
-                    Expr::Binary {
-                        left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                        op: crate::token_kind::TokenKind::FatArrow,
-                        right: Box::new(val_expr),
-                    },
-                ));
-            }
-        }
-    }
-    // :name«words» (French-quote colonpair, equivalent to :name(«words»))
-    if rest.starts_with('\u{00AB}') {
-        let (r, val_expr) = super::container::french_quote_list(rest)?;
-        return Ok((
-            r,
-            Expr::Binary {
-                left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                op: crate::token_kind::TokenKind::FatArrow,
-                right: Box::new(val_expr),
-            },
-        ));
-    }
-    // :name<<words>> (double-angle colonpair, equivalent to :name(<<words>>))
-    if rest.starts_with("<<") {
-        let (r, val_expr) = super::container::double_angle_list(rest)?;
-        return Ok((
-            r,
-            Expr::Binary {
-                left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-                op: crate::token_kind::TokenKind::FatArrow,
-                right: Box::new(val_expr),
-            },
-        ));
-    }
-    // :name (boolean true)
-    Ok((
-        rest,
-        Expr::Binary {
-            left: Box::new(Expr::Literal(Value::str(name.to_string()))),
-            op: crate::token_kind::TokenKind::FatArrow,
-            right: Box::new(Expr::Literal(Value::Bool(true))),
-        },
-    ))
-}
-
-fn render_signature_item(expr: &Expr) -> String {
-    match expr {
-        Expr::BareWord(name) => name.clone(),
-        Expr::Var(name) => format!("${}", name),
-        Expr::ArrayVar(name) => format!("@{}", name),
-        Expr::HashVar(name) => format!("%{}", name),
-        Expr::Binary { left, op, right } if *op == crate::token_kind::TokenKind::FatArrow => {
-            if let Expr::Literal(Value::Str(name)) = left.as_ref() {
-                match right.as_ref() {
-                    Expr::Var(v) if v.as_str() == name.as_str() => format!(":${}", name),
-                    Expr::Literal(Value::Bool(true)) => format!(":${}", name),
-                    other => format!(":${} = {}", name, render_signature_item(other)),
-                }
-            } else {
-                "...".to_string()
-            }
-        }
-        Expr::Literal(v) => match v {
-            Value::Str(s) => format!("\"{}\"", s.replace('"', "\\\"")),
-            _ => v.to_string_value(),
-        },
-        Expr::AnonSub { .. } => "{ ... }".to_string(),
-        _ => "...".to_string(),
-    }
-}
-
-fn parse_signature_fragment(input: &str) -> PResult<'_, String> {
-    let mut depth_paren = 0usize;
-    let mut depth_bracket = 0usize;
-    let mut depth_brace = 0usize;
-    let mut in_single = false;
-    let mut in_double = false;
-    let mut escape = false;
-    for (idx, ch) in input.char_indices() {
-        if in_single {
-            if ch == '\'' && !escape {
-                in_single = false;
-            }
-            escape = ch == '\\' && !escape;
-            continue;
-        }
-        if in_double {
-            if ch == '"' && !escape {
-                in_double = false;
-            }
-            escape = ch == '\\' && !escape;
-            continue;
-        }
-        match ch {
-            '\'' => in_single = true,
-            '"' => in_double = true,
-            '(' => depth_paren += 1,
-            ')' => {
-                if depth_paren == 0 && depth_bracket == 0 && depth_brace == 0 {
-                    let frag = input[..idx].trim();
-                    if frag.is_empty() {
-                        return Err(PError::expected("signature item"));
-                    }
-                    return Ok((&input[idx..], frag.to_string()));
-                }
-                depth_paren = depth_paren.saturating_sub(1);
-            }
-            '[' => depth_bracket += 1,
-            ']' => depth_bracket = depth_bracket.saturating_sub(1),
-            '{' => depth_brace += 1,
-            '}' => depth_brace = depth_brace.saturating_sub(1),
-            ',' if depth_paren == 0 && depth_bracket == 0 && depth_brace == 0 => {
-                let frag = input[..idx].trim();
-                if frag.is_empty() {
-                    return Err(PError::expected("signature item"));
-                }
-                return Ok((&input[idx..], frag.to_string()));
-            }
-            _ => {}
-        }
-    }
-    let frag = input.trim();
-    if frag.is_empty() {
-        return Err(PError::expected("signature item"));
-    }
-    Ok(("", frag.to_string()))
-}
-
-/// Parse capture literals: `\(...)` and `\expr`.
-pub(super) fn capture_literal(input: &str) -> PResult<'_, Expr> {
-    if !input.starts_with('\\') {
-        return Err(PError::expected("capture literal"));
-    }
-    let r = &input[1..]; // skip backslash
-    if !r.starts_with('(') {
-        // `\` as a capture prefix binds tightly — only capture the next
-        // primary term (variable, literal, etc.), not a full expression.
-        // This ensures `\3 eqv \4` parses as `(\3) eqv (\4)`.
-        let (r, item) = super::primary(r)?;
-        return Ok((r, Expr::CaptureLiteral(vec![item])));
-    }
-    // Parenthesized capture literal: \(a, b, :c)
-    let (r, _) = parse_char(r, '(')?;
-    let (r, _) = ws(r)?;
-    let (r, items) = super::parse_call_arg_list(r)?;
-    let (r, _) = ws(r)?;
-    let (r, _) = parse_char(r, ')')?;
-    Ok((r, Expr::CaptureLiteral(items)))
-}
-
-/// Parse `-> $param { body }` or `-> $a, $b { body }` arrow lambda.
-/// Also handles `<-> $a, $b { body }` (rw pointy block) where all params get `is rw`.
-pub(super) fn arrow_lambda(input: &str) -> PResult<'_, Expr> {
-    let is_rw_block = input.starts_with("<->");
-    if !is_rw_block && !input.starts_with("->") {
-        return Err(PError::expected("arrow lambda"));
-    }
-    // Capture the definition line of the pointy block (at the `->` position)
-    let arrow_line = super::current_line_number(input);
-    let arrow_is_original = super::is_within_original_source(input);
-    let r = if is_rw_block {
-        &input[3..]
-    } else {
-        &input[2..]
-    };
-    let (r, _) = ws(r)?;
-    // Zero-param pointed block: -> { body }
-    if r.starts_with('{') {
-        let (r, mut body) = parse_block_body(r)?;
-        if arrow_is_original {
-            body.insert(0, crate::ast::Stmt::SetLine(arrow_line));
-        }
-        return Ok((
-            r,
-            Expr::AnonSubParams {
-                params: Vec::new(),
-                param_defs: Vec::new(),
-                return_type: None,
-                body,
-                is_rw: false,
-                is_whatever_code: false,
-            },
-        ));
-    }
-    // Zero-param with explicit return spec: -> --> 42 { body }
-    if r.starts_with("-->") {
-        let (r, return_type) = skip_pointy_return_type(r)?;
-        let (r, mut body) = parse_block_body(r)?;
-        if arrow_is_original {
-            body.insert(0, crate::ast::Stmt::SetLine(arrow_line));
-        }
-        return Ok((
-            r,
-            Expr::AnonSubParams {
-                params: Vec::new(),
-                param_defs: Vec::new(),
-                return_type,
-                body,
-                is_rw: false,
-                is_whatever_code: false,
-            },
-        ));
-    }
-    // Sub-signature destructuring: -> ($a, $b) { body }.
-    // In Raku, `-> ($a, $b)` is ONE parameter that binds a single list/Capture
-    // argument and unpacks it — NOT two separate parameters (`-> $a, $b`). So a
-    // multi-element parenthesized signature must become a single destructuring
-    // `__subsig__` param (matching `sub f(($a,$b))`), so `.map(-> ($k,$v){...})`
-    // destructures each list element instead of chunking the list 2-at-a-time.
-    // A single-element paren (`-> ($a)`) keeps the existing behavior of a lone
-    // param (mirrors the capture sub-sig `len == 1` special case in sub_param.rs).
-    if r.starts_with('(') {
-        let (r, _) = parse_char(r, '(')?;
-        let (r, _) = ws(r)?;
-        let (r, mut sub_params) = super::super::stmt::parse_param_list_pub(r)?;
-        let (r, _) = ws(r)?;
-        let (r, _) = parse_char(r, ')')?;
-        let (r, return_type) = skip_pointy_return_type(r)?;
-        let (r, mut body) = parse_block_body_with_sigilless(&sub_params, r)?;
-        if arrow_is_original {
-            body.insert(0, crate::ast::Stmt::SetLine(arrow_line));
-        }
-        if is_rw_block {
-            inject_rw_trait(&mut sub_params);
-        }
-        let param_defs = if sub_params.len() > 1 {
-            let mut p = super::super::stmt::sub_param::make_param("__subsig__".to_string());
-            p.sub_signature = Some(sub_params);
-            vec![p]
-        } else {
-            sub_params
-        };
-        let params: Vec<String> = param_defs.iter().map(|p| p.name.clone()).collect();
-        return Ok((
-            r,
-            Expr::AnonSubParams {
-                params,
-                param_defs,
-                return_type,
-                body,
-                is_rw: false,
-                is_whatever_code: false,
-            },
-        ));
-    }
-    // Parse params
-    let (r, first) = super::super::stmt::parse_pointy_param_pub(r)?;
-    let (r, _) = ws(r)?;
-    if r.starts_with(',') {
-        // Multi-param: -> $a, $b { body }
-        let mut param_defs = vec![first];
-        let mut r = r;
-        loop {
-            let (r2, _) = parse_char(r, ',')?;
-            let (r2, _) = ws(r2)?;
-            let (r2, next) = super::super::stmt::parse_pointy_param_pub(r2)?;
-            param_defs.push(next);
-            let (r2, _) = ws(r2)?;
-            if !r2.starts_with(',') {
-                r = r2;
-                break;
-            }
-            r = r2;
-        }
-        let (r, return_type) = skip_pointy_return_type(r)?;
-        let (r, mut body) = parse_block_body_with_sigilless(&param_defs, r)?;
-        if arrow_is_original {
-            body.insert(0, crate::ast::Stmt::SetLine(arrow_line));
-        }
-        if is_rw_block {
-            inject_rw_trait(&mut param_defs);
-        }
-        let params: Vec<String> = param_defs.iter().map(|p| p.name.clone()).collect();
-        Ok((
-            r,
-            Expr::AnonSubParams {
-                params,
-                param_defs,
-                return_type,
-                body,
-                is_rw: false,
-                is_whatever_code: false,
-            },
-        ))
-    } else {
-        // Single param: -> $n { body }
-        let mut first = first;
-        if is_rw_block {
-            inject_rw_trait(std::slice::from_mut(&mut first));
-        }
-        let (r, return_type) = skip_pointy_return_type(r)?;
-        let (r, mut body) = parse_block_body_with_sigilless(std::slice::from_ref(&first), r)?;
-        if arrow_is_original {
-            body.insert(0, crate::ast::Stmt::SetLine(arrow_line));
-        }
-        let simple_single = first.traits.is_empty()
-            && first.shape_constraints.is_none()
-            && !first.named
-            && !first.slurpy
-            && !first.double_slurpy
-            && first.default.is_none()
-            && !first.optional_marker
-            && first.type_constraint.is_none()
-            && first.where_constraint.is_none()
-            && first.sub_signature.is_none()
-            && first.outer_sub_signature.is_none()
-            && first.code_signature.is_none();
-        if simple_single {
-            // Strip sigil prefix for Lambda (it handles sigils internally)
-            let lambda_name = first
-                .name
-                .strip_prefix('@')
-                .or_else(|| first.name.strip_prefix('%'))
-                .or_else(|| first.name.strip_prefix('&'))
-                .unwrap_or(&first.name)
-                .to_string();
-            Ok((
-                r,
-                Expr::Lambda {
-                    param: lambda_name,
-                    body,
-                    is_whatever_code: false,
-                },
-            ))
-        } else {
-            Ok((
-                r,
-                Expr::AnonSubParams {
-                    params: vec![first.name.clone()],
-                    param_defs: vec![first],
-                    return_type,
-                    body,
-                    is_rw: false,
-                    is_whatever_code: false,
-                },
-            ))
-        }
-    }
-}
-
-/// For `<->` blocks, add `is rw` trait to params that don't already have
-/// an explicit `is readonly` trait.
-fn inject_rw_trait(params: &mut [crate::ast::ParamDef]) {
-    for p in params.iter_mut() {
-        if !p.traits.iter().any(|t| t == "readonly") && !p.traits.iter().any(|t| t == "rw") {
-            p.traits.push("rw".to_string());
-        }
-    }
-}
-
-/// Parse a block `{ stmts }` as AnonSub or `{}` / `{ key => val, ... }` as Hash.
-pub(in crate::parser) fn block_or_hash_expr(input: &str) -> PResult<'_, Expr> {
-    if !input.starts_with('{') {
-        return Err(PError::expected("block or hash"));
-    }
-    let r = &input[1..];
-    let (r, _) = ws_inner(r);
-
-    // Empty hash: {}
-    if let Some(rest) = r.strip_prefix('}') {
-        return Ok((rest, Expr::Hash(Vec::new())));
-    }
-
-    // Try to detect if this is a hash literal: { key => val, ... }
-    // Heuristic: if after ws we see `ident =>` or `"str" =>` or `'str' =>`, it's a hash.
-    // However, placeholder variables ($^x, @^x, %^x) force it to be a block.
-    if is_hash_literal_start(r) && !body_has_placeholder_vars(r) {
-        return parse_hash_literal_body(r);
-    }
-
-    // Otherwise parse as a block (anonymous sub)
-    crate::parser::stmt::simple::push_scope();
-    let result = (|| -> PResult<'_, Expr> {
-        let (r, stmts) = super::super::stmt::stmt_list_pub(r)?;
-        let (r, _) = ws_inner(r);
-        if !r.starts_with('}') {
-            return Err(PError::expected("'}'"));
-        }
-        let r = &r[1..];
-        Ok((r, make_anon_sub(stmts)))
-    })();
-    crate::parser::stmt::simple::pop_scope();
-    result
-}
+pub(in crate::parser) use colonpair::colonpair_expr;
+pub(in crate::parser) use lambda::{block_or_hash_expr, parse_block_body};
+pub(in crate::parser) use reduction::reduction_call_style_expr;
 
 /// Simple whitespace consumer that doesn't use PResult (infallible).
 pub(super) fn ws_inner(input: &str) -> (&str, ()) {
-    match super::super::helpers::ws(input) {
+    match crate::parser::helpers::ws(input) {
         Ok((r, _)) => (r, ()),
         Err(_) => (input, ()),
     }
-}
-
-pub(in crate::parser) fn parse_block_body(input: &str) -> PResult<'_, Vec<crate::ast::Stmt>> {
-    let (r, _) = parse_char(input, '{')?;
-    // Reject {YOU_ARE_HERE} outside of a setting (X::Syntax::Reserved)
-    {
-        let trimmed = r.trim_start();
-        if let Some(after) = trimmed.strip_prefix("YOU_ARE_HERE") {
-            let after_trimmed = after.trim_start();
-            if after_trimmed.starts_with('}') {
-                let mut attrs = std::collections::HashMap::new();
-                attrs.insert(
-                    "message".to_string(),
-                    crate::value::Value::str(
-                        "The use of {YOU_ARE_HERE} outside of a setting is reserved".to_string(),
-                    ),
-                );
-                let exc = crate::value::Value::make_instance(
-                    crate::symbol::Symbol::intern("X::Syntax::Reserved"),
-                    attrs,
-                );
-                return Err(PError::fatal_with_exception(
-                    "The use of {YOU_ARE_HERE} outside of a setting is reserved".to_string(),
-                    Box::new(exc),
-                ));
-            }
-        }
-    }
-    crate::parser::stmt::simple::push_scope();
-    let result = (|| -> PResult<'_, Vec<crate::ast::Stmt>> {
-        let (r, stmts) = super::super::stmt::stmt_list_pub(r)?;
-        let (r, _) = ws_inner(r);
-        let (r, _) = parse_char(r, '}')?;
-        Ok((r, stmts))
-    })();
-    crate::parser::stmt::simple::pop_scope();
-    result
-}
-
-/// Like `parse_block_body`, but registers any sigilless parameters as term symbols
-/// before parsing the body.  This prevents bare names like `s` or `q` from being
-/// misinterpreted as substitution / quoting operators inside arrow-lambda bodies.
-fn parse_block_body_with_sigilless<'a>(
-    param_defs: &[crate::ast::ParamDef],
-    input: &'a str,
-) -> PResult<'a, Vec<crate::ast::Stmt>> {
-    let has_sigilless = param_defs.iter().any(|p| p.sigilless);
-    if !has_sigilless {
-        return parse_block_body(input);
-    }
-    let (r, _) = parse_char(input, '{')?;
-    crate::parser::stmt::simple::push_scope();
-    for pd in param_defs {
-        if pd.sigilless {
-            crate::parser::stmt::simple::register_user_term_symbol(&pd.name);
-        }
-    }
-    let result = (|| -> PResult<'_, Vec<crate::ast::Stmt>> {
-        let (r, stmts) = super::super::stmt::stmt_list_pub(r)?;
-        let (r, _) = ws_inner(r);
-        let (r, _) = parse_char(r, '}')?;
-        Ok((r, stmts))
-    })();
-    crate::parser::stmt::simple::pop_scope();
-    result
-}
-
-/// Scan the body (between { and matching }) for placeholder variables ($^x, @^x, %^x).
-/// If any are found, the block should be treated as a Block, not a Hash.
-fn body_has_placeholder_vars(input: &str) -> bool {
-    let mut depth = 1u32;
-    let mut chars = input.chars().peekable();
-    while let Some(c) = chars.next() {
-        match c {
-            '{' => depth += 1,
-            '}' => {
-                depth -= 1;
-                if depth == 0 {
-                    break;
-                }
-            }
-            '$' | '@' | '%' => {
-                if let Some(&'^') = chars.peek() {
-                    chars.next(); // consume '^'
-                    if let Some(&next) = chars.peek()
-                        && (next.is_alphabetic() || next == '_')
-                    {
-                        return true;
-                    }
-                }
-            }
-            // Skip string contents to avoid false positives
-            '\'' => {
-                for sc in chars.by_ref() {
-                    if sc == '\'' {
-                        break;
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    false
-}
-
-/// Check if the input looks like a hash literal start.
-fn is_hash_literal_start(input: &str) -> bool {
-    // %hash variable at start indicates hash literal: {%hash, ...}
-    // Also handles twigil forms like %*dyn, %!attr, %?config, etc.
-    // But {%hash{$_}} or {%hash.foo} is a block containing a hash expression.
-    if let Some(stripped) = input.strip_prefix('%') {
-        // Skip optional twigil (* ! ? ^)
-        let after_twigil = if stripped.starts_with('*')
-            || stripped.starts_with('!')
-            || stripped.starts_with('?')
-            || stripped.starts_with('^')
-        {
-            &stripped[1..]
-        } else {
-            stripped
-        };
-        if let Ok((r, _)) = super::super::stmt::ident_pub(after_twigil) {
-            let (r, _) = ws_inner(r);
-            // If followed by comma, closing brace, or fat arrow, it's a hash literal
-            if r.starts_with(',') || r.starts_with('}') || r.starts_with("=>") {
-                return true;
-            }
-        }
-    }
-    // ident => or "str" => or 'str' =>
-    if let Ok((r, _)) = super::super::stmt::ident_pub(input) {
-        let (r, _) = ws_inner(r);
-        if r.starts_with("=>") {
-            return true;
-        }
-    }
-    // numeric key => val or numeric R=> val (reverse fat arrow, produces a Pair)
-    if let Ok((r, _)) = super::number::integer(input) {
-        let (r, _) = ws_inner(r);
-        if r.starts_with("=>") || r.starts_with("R=>") {
-            return true;
-        }
-    }
-    // Quoted key => val
-    if (input.starts_with('"') || input.starts_with('\''))
-        && let Ok((r, _)) = single_quoted_string(input).or_else(|_| double_quoted_string(input))
-    {
-        let (r, _) = ws_inner(r);
-        if r.starts_with("=>") {
-            return true;
-        }
-    }
-    // Colon pair: :name(expr) or :name or :!name or :Nname — indicates a hash literal
-    if input.starts_with(':') && !input.starts_with("::") {
-        let r = &input[1..];
-        let digit_end = r
-            .char_indices()
-            .take_while(|(_, c)| {
-                crate::builtins::unicode::unicode_decimal_digit_value(*c).is_some()
-            })
-            .last()
-            .map(|(idx, c)| idx + c.len_utf8())
-            .unwrap_or(0);
-        if digit_end > 0 && r[digit_end..].starts_with('<') {
-            return false;
-        }
-        // :!name
-        if r.starts_with('!') && super::super::stmt::ident_pub(&r[1..]).is_ok() {
-            return true;
-        }
-        // :Nname — numeric colon pair (e.g., :1status)
-        if let Some(first) = r.chars().next()
-            && first.is_ascii_digit()
-        {
-            let digit_end = r.find(|c: char| !c.is_ascii_digit()).unwrap_or(r.len());
-            if digit_end < r.len() && super::super::stmt::ident_pub(&r[digit_end..]).is_ok() {
-                return true;
-            }
-        }
-        // :$var / :@var / :%var
-        if (r.starts_with('$') || r.starts_with('@') || r.starts_with('%'))
-            && super::super::stmt::ident_pub(&r[1..]).is_ok()
-        {
-            return true;
-        }
-        // :name or :name(expr) or :name[expr]
-        if let Ok((_r, name)) = super::super::stmt::ident_pub(r)
-            && !matches!(
-                name.as_str(),
-                "my" | "our"
-                    | "has"
-                    | "if"
-                    | "unless"
-                    | "for"
-                    | "while"
-                    | "until"
-                    | "loop"
-                    | "given"
-                    | "when"
-                    | "return"
-            )
-            && !name.starts_with(|c: char| c.is_ascii_digit())
-        {
-            return true;
-        }
-    }
-    false
-}
-
-static ANON_CLASS_COUNTER: AtomicU64 = AtomicU64::new(0);
-static ANON_ROLE_COUNTER: AtomicU64 = AtomicU64::new(0);
-
-fn parse_qualified_ident_with_hyphens<'a>(input: &'a str) -> PResult<'a, String> {
-    let (mut rest, first) = parse_ident_with_hyphens(input)?;
-    let mut full = first.to_string();
-    while let Some(after) = rest.strip_prefix("::") {
-        let (r2, part) = parse_ident_with_hyphens(after)?;
-        full.push_str("::");
-        full.push_str(part);
-        rest = r2;
-    }
-    Ok((rest, full))
-}
-
-/// Parse a class expression: `class { ... }`, `class Foo { ... }`, or `class :: is Parent { ... }`
-/// Named classes in expression context register the class AND return the type object.
-pub(super) fn anon_class_expr(input: &str) -> PResult<'_, Expr> {
-    // Accept optional declarator prefixes used in expression context (e.g. `my class ...`).
-    let input = if let Some(r) = keyword("my", input).or_else(|| keyword("our", input)) {
-        let (r, _) = ws1(r)?;
-        r
-    } else {
-        input
-    };
-    let rest = keyword("class", input).ok_or_else(|| PError::expected("anonymous class"))?;
-    let (rest, _) = ws(rest)?;
-
-    // Accept `class { ... }`, `class :: ...` (anonymous with optional traits),
-    // or `class Name ...` (named class in expression context)
-    let (rest, name, parents, does_roles) = if let Some(r) = rest.strip_prefix("::") {
-        // Skip `::` (anonymous name placeholder)
-        let (r, _) = ws(r)?;
-        // Parse `is Parent` / `does Role` clauses
-        let mut parents = Vec::new();
-        let mut does_roles: Vec<String> = Vec::new();
-        let mut r = r;
-        loop {
-            if let Some(r2) = keyword("is", r) {
-                let (r2, _) = super::super::helpers::ws1(r2)?;
-                let (r2, parent) = parse_qualified_ident_with_hyphens(r2)?;
-                parents.push(parent);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-            } else if let Some(r2) = keyword("does", r) {
-                let (r2, _) = super::super::helpers::ws1(r2)?;
-                let (r2, role) = parse_qualified_ident_with_hyphens(r2)?;
-                parents.push(role.clone());
-                does_roles.push(role);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-            } else {
-                break;
-            }
-        }
-        let id = ANON_CLASS_COUNTER.fetch_add(1, Ordering::Relaxed);
-        (r, format!("__ANON_CLASS_{id}__"), parents, does_roles)
-    } else if rest.starts_with('{') {
-        let id = ANON_CLASS_COUNTER.fetch_add(1, Ordering::Relaxed);
-        (rest, format!("__ANON_CLASS_{id}__"), Vec::new(), Vec::new())
-    } else if rest.starts_with(|c: char| c.is_ascii_uppercase() || c == '_') {
-        // Named class in expression context: `class Foo { ... }`
-        let (r, class_name) = parse_ident_with_hyphens(rest)?;
-        let (r, _) = ws(r)?;
-        // Parse optional `is Parent` / `does Role` clauses
-        let mut parents = Vec::new();
-        let mut does_roles: Vec<String> = Vec::new();
-        let mut r = r;
-        loop {
-            if let Some(r2) = keyword("is", r) {
-                let (r2, _) = super::super::helpers::ws1(r2)?;
-                let (r2, parent) = parse_qualified_ident_with_hyphens(r2)?;
-                parents.push(parent);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-            } else if let Some(r2) = keyword("does", r) {
-                let (r2, _) = super::super::helpers::ws1(r2)?;
-                let (r2, role) = parse_qualified_ident_with_hyphens(r2)?;
-                parents.push(role.clone());
-                does_roles.push(role);
-                let (r2, _) = ws(r2)?;
-                r = r2;
-            } else {
-                break;
-            }
-        }
-        (r, class_name.to_string(), parents, does_roles)
-    } else {
-        return Err(PError::expected("'{' for anonymous class"));
-    };
-
-    if !rest.starts_with('{') {
-        return Err(PError::expected("'{' for anonymous class body"));
-    }
-
-    let (rest, mut body) = parse_block_body(rest)?;
-    // Insert DoesDecl statements at the beginning of the body for `does` clauses
-    for role_name in does_roles.iter().rev() {
-        body.insert(
-            0,
-            Stmt::DoesDecl {
-                name: Symbol::intern(role_name),
-            },
-        );
-    }
-    Ok((
-        rest,
-        Expr::DoStmt(Box::new(Stmt::ClassDecl {
-            name: Symbol::intern(&name),
-            name_expr: None,
-            parents,
-            class_is_rw: false,
-            is_hidden: false,
-            is_lexical: false,
-            hidden_parents: Vec::new(),
-            does_parents: does_roles,
-            repr: None,
-            body,
-            language_version: crate::parser::current_language_version(),
-            custom_traits: Vec::new(),
-            is_unit: false,
-        })),
-    ))
-}
-
-/// Parse an anonymous grammar expression: `grammar { ... }`
-pub(super) fn anon_grammar_expr(input: &str) -> PResult<'_, Expr> {
-    let rest = keyword("grammar", input).ok_or_else(|| PError::expected("anonymous grammar"))?;
-    let (rest, _) = ws(rest)?;
-    if !rest.starts_with('{') {
-        return Err(PError::expected("'{' for anonymous grammar"));
-    }
-    let id = ANON_CLASS_COUNTER.fetch_add(1, Ordering::Relaxed);
-    let name = format!("__ANON_GRAMMAR_{id}__");
-    let (rest, body) = parse_block_body(rest)?;
-    Ok((
-        rest,
-        Expr::DoStmt(Box::new(Stmt::Package {
-            name: Symbol::intern(&name),
-            body,
-            kind: crate::ast::PackageKind::Grammar,
-            is_unit: false,
-            is_my: false,
-        })),
-    ))
-}
-
-/// Parse an anonymous role expression: `role { ... }` or `role :: { ... }`
-pub(super) fn anon_role_expr(input: &str) -> PResult<'_, Expr> {
-    let rest = keyword("role", input).ok_or_else(|| PError::expected("anonymous role"))?;
-    let (rest, _) = ws(rest)?;
-    // Accept optional `::` (null name) before the block
-    let rest = if let Some(r) = rest.strip_prefix("::") {
-        let (r, _) = ws(r)?;
-        r
-    } else {
-        rest
-    };
-    // Must be followed by '{' (no name) to be an anonymous role
-    if !rest.starts_with('{') {
-        return Err(PError::expected("'{' for anonymous role"));
-    }
-    let id = ANON_ROLE_COUNTER.fetch_add(1, Ordering::Relaxed);
-    let name = format!("__ANON_ROLE_{id}__");
-    let (rest, body) = parse_block_body(rest)?;
-    Ok((
-        rest,
-        Expr::DoStmt(Box::new(Stmt::RoleDecl {
-            name: Symbol::intern(&name),
-            type_params: Vec::new(),
-            type_param_defs: Vec::new(),
-            is_export: false,
-            export_tags: Vec::new(),
-            body,
-            is_rw: false,
-            language_version: crate::parser::current_language_version(),
-            custom_traits: Vec::new(),
-        })),
-    ))
-}
-
-/// Parse hash literal body: key => val, :name(val), ... }
-fn parse_hash_literal_body(input: &str) -> PResult<'_, Expr> {
-    let mut pairs = Vec::new();
-    let mut spread_args = Vec::new();
-    let mut rest = input;
-    let mut pending_key: Option<String> = None;
-    loop {
-        let (r, _) = ws_inner(rest);
-        if let Some(rest) = r.strip_prefix('}') {
-            if pending_key.is_some() {
-                return Err(PError::expected("hash value"));
-            }
-            if spread_args.is_empty() {
-                return Ok((rest, Expr::Hash(pairs)));
-            }
-            let mut args = hash_args_from_pairs(pairs);
-            args.append(&mut spread_args);
-            return Ok((
-                rest,
-                Expr::Call {
-                    name: Symbol::intern("hash"),
-                    args,
-                },
-            ));
-        }
-
-        // Try colon pair syntax: :name(expr), :name, :!name, :$var, etc.
-        if r.starts_with(':') && !r.starts_with("::") {
-            if pending_key.is_some() {
-                return Err(PError::expected("hash value"));
-            }
-            let (r_after_pair, (key, mut val)) = parse_colon_pair_entry(r)?;
-            let (r_after_pair, _) = ws_inner(r_after_pair);
-            let r_after_value = if let Some(after_arrow) = r_after_pair.strip_prefix("=>") {
-                let (after_arrow, _) = ws_inner(after_arrow);
-                let (after_arrow, explicit_val) = super::super::expr::expression(after_arrow)?;
-                val = Some(explicit_val);
-                after_arrow
-            } else {
-                r_after_pair
-            };
-            pairs.push((key, val));
-            let (r_after_value, _) = ws_inner(r_after_value);
-            if let Some(stripped) = r_after_value.strip_prefix(',') {
-                rest = stripped;
-            } else if let Some(stripped) = r_after_value.strip_prefix(';') {
-                rest = stripped;
-            } else {
-                rest = r_after_value;
-            }
-            continue;
-        }
-
-        if let Some(key) = pending_key.take() {
-            let (r, val) = super::super::expr::expression(r)?;
-            pairs.push((key, Some(val)));
-            let (r, _) = ws_inner(r);
-            if let Some(stripped) = r.strip_prefix(',') {
-                rest = stripped;
-            } else if let Some(stripped) = r.strip_prefix(';') {
-                rest = stripped;
-            } else {
-                rest = r;
-            }
-            continue;
-        }
-
-        if let Ok((r_key, key)) = parse_simple_hash_key(r) {
-            let (after_key, _) = ws_inner(r_key);
-            if let Some(after_arrow) = after_key.strip_prefix("=>") {
-                let (after_arrow, _) = ws_inner(after_arrow);
-                let (r_val, val) = super::super::expr::expression(after_arrow)?;
-                pairs.push((key, Some(val)));
-                let (r_val, _) = ws_inner(r_val);
-                if let Some(stripped) = r_val.strip_prefix(',') {
-                    rest = stripped;
-                } else if let Some(stripped) = r_val.strip_prefix(';') {
-                    rest = stripped;
-                } else {
-                    rest = r_val;
-                }
-                continue;
-            }
-
-            // If followed by R=> (reverse fat arrow meta-operator), don't treat as
-            // simple key — fall through to expression parsing which handles R=> correctly.
-            if !after_key.starts_with("R=>") {
-                pending_key = Some(key);
-                let (r_key, _) = ws_inner(r_key);
-                if let Some(stripped) = r_key.strip_prefix(',') {
-                    rest = stripped;
-                } else if let Some(stripped) = r_key.strip_prefix(';') {
-                    rest = stripped;
-                } else {
-                    rest = r_key;
-                }
-                continue;
-            }
-        }
-
-        let (r, item) = super::super::expr::expression(r)?;
-        match item {
-            Expr::HashVar(name) => {
-                spread_args.push(Expr::Unary {
-                    op: crate::token_kind::TokenKind::Pipe,
-                    expr: Box::new(Expr::HashVar(name)),
-                });
-            }
-            Expr::Unary {
-                op: crate::token_kind::TokenKind::Pipe,
-                expr,
-            } => {
-                spread_args.push(Expr::Unary {
-                    op: crate::token_kind::TokenKind::Pipe,
-                    expr,
-                });
-            }
-            other => {
-                if let Some((key, val)) = hash_pair_from_expr(&other)? {
-                    pairs.push((key, Some(val)));
-                } else if let Expr::ArrayLiteral(elems) = &other {
-                    // Flatten word lists like <a b c d> into pairs: a => "b", c => "d"
-                    let mut i = 0;
-                    while i + 1 < elems.len() {
-                        let key = hash_key_from_expr(elems[i].clone())?;
-                        pairs.push((key, Some(elems[i + 1].clone())));
-                        i += 2;
-                    }
-                    if i < elems.len() {
-                        // Odd element becomes a pending key
-                        pending_key = Some(hash_key_from_expr(elems[i].clone())?);
-                    }
-                } else if matches!(other, Expr::Literal(_) | Expr::BareWord(_)) {
-                    pending_key = Some(hash_key_from_expr(other)?);
-                } else {
-                    // A complex expression that is neither a `=>` pair nor a
-                    // simple key (e.g. a ternary `$x ?? :$x !! ()` used as a
-                    // hash element) is a BARE element: raku flattens it into the
-                    // initializer (warning about odd counts only at runtime).
-                    // Route it through the hash() builder via spread_args instead
-                    // of failing with "expected hash key".
-                    spread_args.push(other);
-                }
-            }
-        }
-
-        let (r, _) = ws_inner(r);
-        if let Some(stripped) = r.strip_prefix(',') {
-            rest = stripped;
-        } else if let Some(stripped) = r.strip_prefix(';') {
-            rest = stripped;
-        } else {
-            rest = r;
-        }
-    }
-}
-
-/// Parse the body of an object hash `:{ ... }`.
-/// Object hashes allow arbitrary expression keys (e.g., regex), so we parse
-/// each entry using full expression parsing and emit a `hash(...)` call with
-/// Pair arguments, preserving expression-typed keys at runtime.
-fn parse_object_hash_body(input: &str) -> PResult<'_, Expr> {
-    let mut args = Vec::new();
-    let mut rest = input;
-    loop {
-        let (r, _) = ws_inner(rest);
-        if let Some(rest) = r.strip_prefix('}') {
-            return Ok((
-                rest,
-                Expr::Call {
-                    name: crate::symbol::Symbol::intern("hash"),
-                    args,
-                },
-            ));
-        }
-
-        // Parse each entry as a full expression (which handles `expr => expr` as
-        // a fat arrow binary). This allows regex keys like `rx/.../ => "..."`.
-        let (r, item) = super::super::expr::expression(r)?;
-        args.push(item);
-
-        let (r, _) = ws_inner(r);
-        if let Some(stripped) = r.strip_prefix(',') {
-            rest = stripped;
-        } else if let Some(stripped) = r.strip_prefix(';') {
-            rest = stripped;
-        } else {
-            rest = r;
-        }
-    }
-}
-
-fn hash_args_from_pairs(pairs: Vec<(String, Option<Expr>)>) -> Vec<Expr> {
-    pairs
-        .into_iter()
-        .map(|(key, val_opt)| Expr::Binary {
-            left: Box::new(Expr::Literal(Value::str(key))),
-            op: crate::token_kind::TokenKind::FatArrow,
-            right: Box::new(val_opt.unwrap_or(Expr::Literal(Value::Nil))),
-        })
-        .collect()
-}
-
-fn parse_simple_hash_key(input: &str) -> PResult<'_, String> {
-    if let Ok((r, name)) = super::super::stmt::ident_pub(input) {
-        return Ok((r, name));
-    }
-    if let Ok((r, Expr::Literal(Value::Int(n)))) = super::number::integer(input) {
-        return Ok((r, n.to_string()));
-    }
-    if let Ok((r, Expr::Literal(Value::BigInt(n)))) = super::number::integer(input) {
-        return Ok((r, n.to_string()));
-    }
-    if let Ok((r, Expr::Literal(Value::Str(s)))) =
-        single_quoted_string(input).or_else(|_| double_quoted_string(input))
-    {
-        return Ok((r, s.to_string()));
-    }
-    Err(PError::expected("hash key"))
-}
-
-fn hash_key_from_expr(expr: Expr) -> Result<String, PError> {
-    match expr {
-        // Hash storage currently uses string keys; accept literal keys broadly and
-        // normalize using Raku stringification semantics.
-        Expr::Literal(v) => Ok(v.to_string_value()),
-        Expr::BareWord(name) => Ok(name),
-        _ => Err(PError::expected("hash key")),
-    }
-}
-
-fn hash_pair_from_expr(expr: &Expr) -> Result<Option<(String, Expr)>, PError> {
-    match expr {
-        Expr::Binary {
-            left,
-            op: crate::token_kind::TokenKind::FatArrow,
-            right,
-        } => Ok(Some((
-            hash_key_from_expr((**left).clone())?,
-            (**right).clone(),
-        ))),
-        // R=> (reverse fat arrow): `1 R=> "a"` → key="a", value=1
-        Expr::MetaOp {
-            meta,
-            op,
-            left,
-            right,
-        } if meta == "R" && op == "=>" => Ok(Some((
-            hash_key_from_expr((**right).clone())?,
-            (**left).clone(),
-        ))),
-        Expr::PositionalPair(inner) => hash_pair_from_expr(inner),
-        _ => Ok(None),
-    }
-}
-
-/// Parse a colon pair entry inside a hash literal: :name(expr), :name, :!name, :Nname
-fn parse_colon_pair_entry(input: &str) -> PResult<'_, (String, Option<Expr>)> {
-    let r = input
-        .strip_prefix(':')
-        .ok_or_else(|| PError::expected("':'"))?;
-    let digit_end = r
-        .char_indices()
-        .take_while(|(_, c)| crate::builtins::unicode::unicode_decimal_digit_value(*c).is_some())
-        .last()
-        .map(|(idx, c)| idx + c.len_utf8())
-        .unwrap_or(0);
-    if digit_end > 0 && r[digit_end..].starts_with('<') {
-        return Err(PError::expected("generic radix literal"));
-    }
-
-    // :!name
-    if let Some(r) = r.strip_prefix('!') {
-        let (r, name) = super::super::stmt::ident_pub(r)?;
-        return Ok((r, (name, Some(Expr::Literal(Value::Bool(false))))));
-    }
-
-    // :Nname — numeric colon pair, e.g., :1status means status => 1
-    if let Some(first) = r.chars().next()
-        && first.is_ascii_digit()
-    {
-        let digit_end = r.find(|c: char| !c.is_ascii_digit()).unwrap_or(r.len());
-        let digits = &r[..digit_end];
-        let after_digits = &r[digit_end..];
-        if let Ok((r, name)) = super::super::stmt::ident_pub(after_digits) {
-            let num: i64 = digits.parse().unwrap_or(0);
-            return Ok((r, (name, Some(Expr::Literal(Value::Int(num))))));
-        }
-    }
-
-    // :$var / :@var / :%var / :&var (with twigil support)
-    if r.starts_with('$') || r.starts_with('@') || r.starts_with('%') || r.starts_with('&') {
-        let sigil = &r[..1];
-        let after_sigil = &r[1..];
-        // Check for twigils: *, ?, ^, =, ~, :, !, .
-        let (after_twigil, twigil) = if after_sigil.starts_with('*')
-            || after_sigil.starts_with('?')
-            || after_sigil.starts_with('^')
-            || after_sigil.starts_with('=')
-            || after_sigil.starts_with('~')
-            || after_sigil.starts_with('!')
-            || after_sigil.starts_with('.')
-        {
-            (&after_sigil[1..], &after_sigil[..1])
-        } else if after_sigil.starts_with(':') && !after_sigil.starts_with("::") {
-            (&after_sigil[1..], ":")
-        } else {
-            (after_sigil, "")
-        };
-        let (r, name) = super::super::stmt::ident_pub(after_twigil)?;
-        let full_var_name = if twigil.is_empty() {
-            name.clone()
-        } else {
-            format!("{}{}", twigil, name)
-        };
-        let expr = match sigil {
-            "$" => Expr::Var(full_var_name),
-            "@" => Expr::ArrayVar(full_var_name),
-            "%" => Expr::HashVar(full_var_name),
-            "&" => Expr::CodeVar(full_var_name),
-            _ => unreachable!(),
-        };
-        return Ok((r, (name, Some(expr))));
-    }
-
-    // :name or :name(expr) or :name[items]
-    let (r, name) = super::super::stmt::ident_pub(r)?;
-
-    // :name(expr) or :name(expr, expr, ...)
-    if r.starts_with('(') {
-        let (r, _) = parse_char(r, '(')?;
-        let (r, _) = ws_inner(r);
-        let (r, first) = super::super::expr::expression(r)?;
-        let (r, _) = ws_inner(r);
-        if r.starts_with(',') {
-            let mut items = vec![first];
-            let mut r = r;
-            while r.starts_with(',') {
-                let (r2, _) = parse_char(r, ',')?;
-                let (r2, _) = ws_inner(r2);
-                let (r2, next) = super::super::expr::expression(r2)?;
-                let (r2, _) = ws_inner(r2);
-                items.push(next);
-                r = r2;
-            }
-            let (r, _) = parse_char(r, ')')?;
-            return Ok((r, (name, Some(Expr::ArrayLiteral(items)))));
-        }
-        let (r, _) = parse_char(r, ')')?;
-        return Ok((r, (name, Some(first))));
-    }
-
-    // :name[items]
-    if r.starts_with('[') {
-        let (r, _) = parse_char(r, '[')?;
-        let (r, _) = ws_inner(r);
-        let mut items = Vec::new();
-        let mut r = r;
-        while !r.starts_with(']') {
-            let (r2, item) = super::super::expr::expression(r)?;
-            items.push(item);
-            let (r2, _) = ws_inner(r2);
-            if r2.starts_with(',') {
-                let (r2, _) = parse_char(r2, ',')?;
-                let (r2, _) = ws_inner(r2);
-                r = r2;
-            } else {
-                r = r2;
-            }
-        }
-        let (r, _) = parse_char(r, ']')?;
-        return Ok((r, (name, Some(Expr::BracketArray(items, false)))));
-    }
-
-    // :name{...} (block/hash-valued colonpair)
-    if r.starts_with('{') {
-        let (r, block_or_hash) = block_or_hash_expr(r)?;
-        return Ok((r, (name, Some(block_or_hash))));
-    }
-
-    // :name<word> or :name<words> (angle bracket form)
-    if r.starts_with('<') && !r.starts_with("<<") {
-        let (r, _) = parse_char(r, '<')?;
-        let end = r
-            .find('>')
-            .ok_or_else(|| PError::expected("'>' closing angle bracket"))?;
-        let content = &r[..end];
-        let r = &r[end + 1..];
-        let words = split_angle_words(content);
-        if words.len() == 1 {
-            return Ok((
-                r,
-                (name, Some(Expr::Literal(Value::str(words[0].to_string())))),
-            ));
-        }
-        let items = words
-            .iter()
-            .map(|w| Expr::Literal(Value::str(w.to_string())))
-            .collect();
-        return Ok((r, (name, Some(Expr::ArrayLiteral(items)))));
-    }
-
-    // :name«words» (French-quote form)
-    if r.starts_with('\u{00AB}') {
-        let (r, val_expr) = super::container::french_quote_list(r)?;
-        return Ok((r, (name, Some(val_expr))));
-    }
-
-    // :name<<words>> (double-angle form)
-    if r.starts_with("<<") {
-        let (r, val_expr) = super::container::double_angle_list(r)?;
-        return Ok((r, (name, Some(val_expr))));
-    }
-
-    // :name (boolean true)
-    Ok((r, (name, None)))
 }
